Refactored Ping Request and Ping Response to be implemented by a single DuplexChannelHandler
