--- hivemq-extension-sdk/src/main/java/com/hivemq/extension/sdk/api/interceptor/pingrequestresponse/parameter/PingRequestResponseInput.java
@@ -1,6 +1,0 @@
-package com.hivemq.extension.sdk.api.interceptor.pingrequestresponse.parameter;
-
-import com.hivemq.extension.sdk.api.parameter.ClientBasedInput;
-
-public interface PingRequestResponseInput extends ClientBasedInput {
-}
