--- hivemq-extension-sdk/src/main/java/com/hivemq/extension/sdk/api/interceptor/pingrequestresponse/parameter/PingRequestResponseOutput.java
@@ -1,6 +1,0 @@
-package com.hivemq.extension.sdk.api.interceptor.pingrequestresponse.parameter;
-
-import com.hivemq.extension.sdk.api.async.AsyncOutput;
-
-public interface PingRequestResponseOutput extends AsyncOutput<PingRequestResponseOutput> {
-}
