+++ src/main/java/com/hivemq/extensions/client/ClientContextPluginImpl.java
@@ -21,7 +21,6 @@ import com.hivemq.annotations.NotNull;
 import com.hivemq.extension.sdk.api.client.ClientContext;
 import com.hivemq.extension.sdk.api.interceptor.Interceptor;
 import com.hivemq.extension.sdk.api.interceptor.pingrequest.PingRequestInboundInterceptor;
-import com.hivemq.extension.sdk.api.interceptor.pingrequestresponse.PingRequestResponseInterceptor;
 import com.hivemq.extension.sdk.api.interceptor.pingresponse.PingResponseOutboundInterceptor;
 import com.hivemq.extension.sdk.api.interceptor.publish.PublishInboundInterceptor;
 import com.hivemq.extension.sdk.api.interceptor.publish.PublishOutboundInterceptor;
@@ -78,12 +77,6 @@ public class ClientContextPluginImpl extends AbstractOutput implements ClientCon
     }
 
     @Override
-    public void addPingRequestResponseInterceptor(
-            @NotNull final PingRequestResponseInterceptor interceptor) {
-        clientContext.removeInterceptor(checkNotNull(interceptor, "The interceptor must never be null"));
-    }
-
-    @Override
     public void removePublishInboundInterceptor(final @NotNull PublishInboundInterceptor interceptor) {
         clientContext.removeInterceptor(checkNotNull(interceptor, "The interceptor must never be null"));
     }
@@ -108,11 +101,6 @@ public class ClientContextPluginImpl extends AbstractOutput implements ClientCon
         clientContext.removeInterceptor(checkNotNull(interceptor, "The interceptor must never be null"));
     }
 
-    @Override
-    public void removePingRequestResponseInterceptor(
-            @NotNull final PingRequestResponseInterceptor interceptor) {
-        clientContext.removeInterceptor(checkNotNull(interceptor, "The interceptor must never be null"));
-    }
 
     @NotNull
     @Override
